dom/notification/test/mochitest/test_notificationclick.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=916893
-->
<head>
  <title>Bug 1114554 - Test ServiceWorkerGlobalScope.notificationclick event.</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/dom/notification/test/mochitest/MockAlertsService.js"></script>
  <script src="/tests/dom/notification/test/mochitest/NotificationTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1114554">Bug 1114554</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
<script src="/tests/dom/serviceworkers/test/utils.js"></script>
<script>
  SimpleTest.requestFlakyTimeout("Mock alert service dispatches show and click events.");

  async function testFrame(src, registration) {
    let { promise, resolve } = Promise.withResolvers();
    let iframe = document.createElement("iframe");
    iframe.src = src;
    window.callback = async function(result) {
      window.callback = null;
      document.body.removeChild(iframe);
      iframe = null;
      ok(result, "Got notificationclick event with correct data.");
      await registration.unregister();
      resolve();
    };
    document.body.appendChild(iframe);
    await promise;
  }

  add_task(async function runTest() {
    await NotificationTest.allowNotifications();
    await MockAlertsService.register();
    await MockAlertsService.enableAutoClick();
    let registration = await navigator.serviceWorker.register(
      "notificationclick.js",
      { scope: "notificationclick.html" }
    );
    await waitForState(registration.installing, 'activated');
    // Now that we know the document will be controlled, create the frame.
    await testFrame('notificationclick.html', registration);
  });
</script>
</body>
</html>
